1.1.0 Update: Added WorldManagement in Wrapper & CLoudPlugin (CloudWorld, WorldProperties, WorldDownloader), improved JenkinsDownloader, added MasterServer reload command (for Templates and static Servers), changed Template concept (no more zip files -> new properties json String in database), added new ConfigSetter class (for modifying templates from within the template, before server starts), improved BungeeCord ServerList Updates (merged 2 packets), added config.yml for Wrapper (for dynamic jenkinsPlugin & world caching -> with build number), fixed Decoder for empty ByteBufs, removed OS class, added CloudException for WrapperDownloader >> pre-ready for live system

+++ wrapper/src/main/java/eu/mcone/cloud/wrapper/server/Server.java
@@ -5,28 +5,34 @@
 
 package eu.mcone.cloud.wrapper.server;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
 import eu.mcone.cloud.core.console.Logger;
+import eu.mcone.cloud.core.exception.CloudException;
 import eu.mcone.cloud.core.file.UnZip;
-import eu.mcone.cloud.core.network.packet.ServerResultPacketWrapper;
 import eu.mcone.cloud.core.network.packet.ServerUpdateStatePacket;
 import eu.mcone.cloud.core.server.ServerInfo;
 import eu.mcone.cloud.core.server.ServerState;
+import eu.mcone.cloud.core.server.world.CloudWorld;
 import eu.mcone.cloud.wrapper.WrapperServer;
-import eu.mcone.cloud.wrapper.jenkins.JenkinsDownloader;
+import eu.mcone.cloud.wrapper.download.JenkinsDownloader;
+import eu.mcone.cloud.wrapper.download.WorldDownloader;
 import eu.mcone.cloud.wrapper.server.console.ConsoleInputReader;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.commons.io.FileUtils;
 
-import java.io.*;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
 import java.lang.reflect.InvocationTargetException;
 import java.net.Socket;
-import java.net.URL;
-import java.nio.channels.Channels;
 
 public abstract class Server {
 
     private static final File homeDir = WrapperServer.getInstance().getFileManager().getHomeDir();
+    protected Gson gson;
 
     @Getter @Setter
     protected ServerInfo info;
@@ -39,10 +45,14 @@ public abstract class Server {
     @Getter
     protected ServerState state = ServerState.OFFLINE;
     @Getter
+    protected ServerProperties properties;
+    @Getter
     private ConsoleInputReader reader;
 
     public Server(ServerInfo info) {
+        this.gson = new Gson();
         this.info = info;
+        this.properties = gson.fromJson(info.getProperties(), ServerProperties.class);
 
         if (info.isStaticServer()) {
             this.serverDir = new File(homeDir + File.separator + "staticservers" + File.separator + info.getName());
@@ -60,41 +70,66 @@ public abstract class Server {
     abstract void setConfig() throws IOException;
 
     void initialise(final int port, final Class<? extends ConsoleInputReader> reader, final String[] command) {
+        if (!state.equals(ServerState.OFFLINE)) {
+            Logger.err(getClass(), "["+info.getName()+"] Cannot start Server if not OFFLINE");
+            return;
+        }
+
         info.setPort(port);
 
         Logger.log(getClass(), "["+info.getName()+"] Starting server (Version: '"+info.getVersion()+"', UUID: '" + info.getUuid() + "', Template: '" + info.getTemplateName() + "', RAM: '" + info.getRam() + "M, Port: '" + info.getPort() + "')...");
         setState(ServerState.STARTING);
 
         WrapperServer.getInstance().getThreadPool().execute(() -> {
-            final File templateZip = new File(serverDir+File.separator+info.getTemplateName()+".zip");
             final File executable = new File(homeDir+File.separator+"jars"+File.separator+info.getVersion().toString()+".jar");
+            final File worldFile = new File(serverDir+File.separator+"worlds.json");
 
             try {
                 if (!info.isStaticServer()) {
                     if (serverDir.exists()) FileUtils.deleteDirectory(serverDir);
                     serverDir.mkdir();
 
-                    Logger.log(getClass(), "[" + info.getName() + "] Downloading Template...");
-                    URL website = new URL("http://templates.mcone.eu/" + info.getTemplateName() + ".zip");
-                    FileOutputStream fos = new FileOutputStream(templateZip);
-                    fos.getChannel().transferFrom(Channels.newChannel(website.openStream()), 0, Long.MAX_VALUE);
+                    for (ServerProperties.PluginDownload download : properties.getPlugins()) {
+                        File plugin = new JenkinsDownloader(JenkinsDownloader.CiServer.valueOf(download.getCiServer())).getJenkinsArtifact(download.getJob(), download.getArtifact());
 
-                    Logger.log(getClass(), "[" + info.getName() + "] Unzipping Template...");
-                    new UnZip(templateZip.getPath(), serverDir.getPath());
-                    templateZip.delete();
+                        Logger.log(getClass(), "["+info.getName()+"] Implementing Plugin "+download.getJob()+":"+download.getArtifact());
+                        FileUtils.copyFile(
+                                plugin,
+                                new File(serverDir + File.separator + "plugins" + File.separator + plugin.getName())
+                        );
+                    }
+
+                    JsonArray worldArray = new JsonArray();
+                    for (String w : properties.getWorlds()) {
+                        CloudWorld world = new WorldDownloader(w).downloadWorld();
+                        worldArray.add(gson.toJsonTree(world, CloudWorld.class));
+
+                        Logger.log(getClass(), "["+info.getName()+"] Implementing World "+w);
+                        new UnZip(
+                                world.getFilePath(),
+                                serverDir.getPath() + File.separator + w
+                        );
+                    }
+
+                    worldFile.createNewFile();
+                    FileUtils.writeStringToFile(worldFile, gson.toJson(worldArray));
                 } else {
                     if (!serverDir.exists()) serverDir.mkdir();
                 }
 
-                Logger.log(getClass(), "Implementing Jenkins Files");
+                Logger.log(getClass(), "Implementing Cloud-Plugin");
                 FileUtils.copyFile(
-                        new JenkinsDownloader(JenkinsDownloader.CiServer.MCONE).getJenkinsArtifact("MCONE-Cloud", "plugin"),
+                        //new JenkinsDownloader(JenkinsDownloader.CiServer.MCONE).getJenkinsArtifact("MCONE-Cloud", "plugin"),
+                        new File("D:\\Rufus Maiwald\\Documents\\Java\\Projekte\\mc1cloud\\out\\artifacts\\plugin_jar\\plugin.jar"),
                         new File(serverDir + File.separator + "plugins" + File.separator + "MCONE-CloudPlugin.jar")
                 );
 
                 FileUtils.copyFile(executable, new File(serverDir+File.separator+"server.jar"));
 
                 setConfig();
+                for (ServerProperties.Config config : properties.getConfigs()) {
+                    new ConfigSetter(this, config);
+                }
 
                 this.runtime = Runtime.getRuntime();
                 this.process = this.runtime.exec(command, null, serverDir);
@@ -112,13 +147,13 @@ public abstract class Server {
                 } else if (state.equals(ServerState.STARTING)) {
                     Logger.err(getClass(), "["+info.getName()+"] Server crashed while starting! Fix this problem before starting it again!");
                 }
-            } catch (IOException | InterruptedException | IllegalAccessException | InstantiationException | InvocationTargetException | NoSuchMethodException e) {
+            } catch (IOException | InterruptedException | IllegalAccessException | InstantiationException | InvocationTargetException | NoSuchMethodException | CloudException e) {
                 Logger.log(getClass(), "["+info.getName()+"] Could not start server:");
-                if (e instanceof FileNotFoundException) {
-                    Logger.err(getClass(), "["+info.getName()+"] Template does not exist, cancelling...");
-                    e.printStackTrace();
+                if (e instanceof CloudException) {
+                    Logger.err(getClass(), "["+info.getName()+"] "+e.getMessage());
                     return;
                 }
+
                 e.printStackTrace();
             }
         });
@@ -136,14 +171,11 @@ public abstract class Server {
                 Logger.log(getClass(), "["+info.getName()+"] ForceStop server " + info.getName() + "...");
                 process.destroy();
                 setState(ServerState.OFFLINE);
-                sendResult("[Server." + this.info.getName() + "] The server was ForceStopped...", ServerResultPacketWrapper.Result.SUCCESSFUL);
             } else {
                 Logger.log(getClass(), "["+info.getName()+"] Could not be forcestop server because the process is dead...");
-                this.sendResult("[Server." + this.info.getName() + "] The server cloud not be ForeStopped because the process is dead...", ServerResultPacketWrapper.Result.COOMMAND_ERROR);
             }
         } else {
             Logger.log(getClass(), "["+info.getName()+"] Could not forcestop because server has no process...");
-            this.sendResult("[Server." + this.info.getName() + "] The server could not be ForeStopped because it has no process...", ServerResultPacketWrapper.Result.COOMMAND_ERROR);
         }
     }
 
@@ -155,11 +187,6 @@ public abstract class Server {
         WrapperServer.getInstance().getServers().remove(this);
 
         Logger.log(getClass(), "["+info.getName()+"] Server deleted...");
-        this.sendResult("[Server." + server_name + "] The server was deleted...", ServerResultPacketWrapper.Result.SUCCESSFUL);
-    }
-
-    private boolean isAlive() {
-        return process.isAlive();
     }
 
     public void sendCommand(String command) {
@@ -171,18 +198,14 @@ public abstract class Server {
                     out.flush();
 
                     Logger.log(getClass(), "["+info.getName()+"] Sent command '" + command + "'");
-                    this.sendResult("[Server." + this.info.getName() + "] Sent command '" + command + "' to server...", ServerResultPacketWrapper.Result.COMMAND);
                 } else {
                     Logger.err(getClass(), "["+info.getName()+"] Could not send command '" + command + "' because the process is dead...");
-                    this.sendResult("[Server." + this.info.getName() + "] The command '" + command + "' could not be sent to server '" + this.info.getName() + "' because the process is dead...", ServerResultPacketWrapper.Result.COOMMAND_ERROR);
                 }
             } else {
                 Logger.err(getClass(), "["+info.getName()+"] Could not send command '" + command + "' because server has no process...");
-                this.sendResult("[Server." + this.info.getName() + "] The command '" + command + "' could not be sent to server because it has no process...", ServerResultPacketWrapper.Result.COOMMAND_ERROR);
             }
         } catch (IOException e) {
             Logger.err(getClass(), "["+info.getName()+"] Could not send command '" + command + "' to the server (IOException)");
-            this.sendResult("[Server." + this.info.getName() + "] the command '" + command + "' cloud not be sent to server...", ServerResultPacketWrapper.Result.COOMMAND_ERROR);
             e.printStackTrace();
         }
     }
@@ -190,11 +213,6 @@ public abstract class Server {
     public void setState(ServerState state) {
         this.state = state;
         WrapperServer.getInstance().send(new ServerUpdateStatePacket(info.getUuid(), state));
-    }
-
-    void sendResult(String message, ServerResultPacketWrapper.Result result) {
-        WrapperServer.getInstance().send(new ServerResultPacketWrapper("Server.class", message, result));
-        System.out.println("[Server.class] The result '" + message + "\\" + result.toString() + "' was sent to the master...");
     }
 
     private static int getNextAvailablePort(int port) {
